requirements.txt
absl-py==1.4.0
accelerate==0.17.1
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
async-timeout==4.0.2
attrs==22.2.0
blessed==1.20.0
brotlipy==0.7.0
certifi @ file:///opt/conda/conda-bld/certifi_1655968806487/work/certifi
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
datasets==2.10.1
deepspeed==0.8.2
dill==0.3.6
docker-pycreds==0.4.0
evaluate==0.4.0
filelock==3.9.0
fonttools==4.25.0
frozenlist==1.3.3
fsspec==2023.3.0
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
gpustat==1.0.0
hjson==3.1.0
huggingface-hub==0.13.2
idna==3.4
joblib==1.2.0
jsonline==0.2.1
jsonlines==3.1.0
loguru==0.6.0
lxml==4.9.2
multidict==6.0.4
multiprocess==0.70.14
munkres==1.1.4
ninja==1.11.1
nltk==3.8.1
numpy==1.24.2
nvidia-ml-py==11.495.46
omegaconf==2.3.0
packaging==23.0
pandas==1.5.3
parse==1.19.0
pathtools==0.1.2
peft==0.3.0
ply==3.11
portalocker==2.7.0
protobuf==4.22.1
psutil==5.9.4
py-cpuinfo==9.0.0
pyarrow==11.0.0
pycountry==22.3.5
pycparser==2.21
pydantic==1.10.6
pynvml==11.5.0
PyQt5-sip==12.11.0
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
responses==0.18.0
rouge-score==0.1.2
sacrebleu==2.3.1
scikit-learn==1.2.2
scipy==1.9.3
sentencepiece==0.1.97
sentry-sdk==1.16.0
setproctitle==1.3.2
six==1.16.0
sklearn==0.0.post1
smmap==5.0.0
sqlitedict==2.1.0
tabulate==0.9.0
threadpoolctl==3.1.0
tokenizers==0.13.2
torch==1.13.1+cu117
tqdm==4.65.0
transformers @ git+https://github.com/zphang/transformers@ef61b1ba1a8ee9fd354b640b059c3474b676c0c5
typing_extensions==4.5.0
urllib3==1.26.15
wandb==0.14.0
wcwidth==0.2.6
xxhash==3.2.0
yarl==1.8.2
